for sharing
