+++ src/features/suggestions/search/PredefinedSuggestion.ts
@@ -1,0 +1,9 @@
+import { Serving } from "../../../model/Food";
+
+
+export interface PredefinedSuggestion {
+  foodName: string;
+  amount: string;
+  serving: Serving;
+  bestChoice?: boolean;
+}
